requirements_dev.txt
Django==4.1.1
autopep8==1.7.0
django-parler==2.3
django-parler-rest==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
drf-yasg[validation]==1.21.3
Faker==14.2.0
sorl-thumbnail==12.9.
sorl-thumbnail-serializer-field==0.2.1
django-cleanup==6.0.0
Pillow==9.2.0
django-enumchoicefield==3.0.1
psycopg2-binary==2.9.3
uvicorn==0.18.3
gunicorn==20.1.0
django-environ==0.9.0
django-storages[google]==1.13.1